requirements.txt
gradio>=4.0
httpx
Pillow
rembg
numpy
python-dotenv
